グラフをレスポンシブにする
